added content descriptions to intro tab; moved profile tab functions to new script; updated readme file; updated environemnt.yml and requirements.txt files
