+++ modules/MCTab.py
@@ -1,8 +1,8 @@
-import os
-import requests
-import json
+# import os
+# import requests
+# import json
 import pandas as pd
-from dotenv import load_dotenv
+# from dotenv import load_dotenv
 
 from modules.MCForecastTools import MCSimulation
 import warnings
