+++ modules/intro.py
@@ -2,6 +2,7 @@
 
 def get_intro():
     text = """\
+    ## Welcome to the MyWealthPlan Portfolio Selection Tool
     
     ### To begin, please complete the Risk Analysis Survey found in the left-hand side panel. Once you have selected your answers click on the 'submit' button.
 
@@ -12,7 +13,11 @@ def get_intro():
     ### You may naviate through the informational displays by selecting the desired tab from the menu above. The information is organized as follows:
 
     ----
-
+   
+    * "Portfolio Profile" provides information on the allocation of assets between Stocks, Bonds, Cryptocurrencies and Commodoties
+    * "Past Performance" provides information regarding the historical performance of the selected portfolio along with a comparison to the historical performance of the S&P 500
+    * "Future Performance" provides the results of a Monte Carlo simulation using 200 iterations and can be useful to gauge the potential future performance of the portfolio
+   
     """
     
     return text
@@ -22,7 +27,7 @@ def get_portfolios_intro():
     text = """\
     
     ----
-    ### Based on your answers, you will be matched with one of those five portfolios: 
+    ### Based on your answers, you will be matched with one of these five portfolios: 
 
     """
