+++ requirements.txt
@@ -1,127 +1,8 @@
-anyio==3.6.2
-appdirs==1.4.4
-appnope==0.1.3
-argon2-cffi==21.3.0
-argon2-cffi-bindings==21.2.0
-arrow==1.2.3
-asttokens==2.2.1
-attrs==22.2.0
-Babel==2.11.0
-backcall==0.2.0
-beautifulsoup4==4.11.2
-bleach==6.0.0
 bokeh==2.4.3
-bs4==0.0.1
-certifi==2022.12.7
-cffi==1.15.1
-charset-normalizer==3.0.1
-colorcet==3.0.1
-comm==0.1.2
-contourpy==1.0.7
-cssselect==1.2.0
-cycler==0.11.0
-debugpy==1.6.6
-decorator==5.1.1
-defusedxml==0.7.1
-executing==1.2.0
-fake-useragent==1.1.1
-fastjsonschema==2.16.2
-feedparser==6.0.10
-fonttools==4.38.0
-fqdn==1.5.1
-holoviews==1.15.4
 hvplot==0.8.2
-idna==3.4
-importlib-metadata==6.0.0
-importlib-resources==5.10.2
-ipykernel==6.21.2
-ipython==8.10.0
-ipython-genutils==0.2.0
-isoduration==20.11.0
-jedi==0.18.2
-Jinja2==3.1.2
-json5==0.9.11
-jsonpointer==2.3
-jsonschema==4.17.3
-jupyter-events==0.6.3
-jupyter_client==8.0.2
-jupyter_core==5.2.0
-jupyter_server==2.2.1
-jupyter_server_terminals==0.4.4
 jupyterlab==3.5.0
-jupyterlab-pygments==0.2.2
-jupyterlab_server==2.19.0
-kiwisolver==1.4.4
-lxml==4.9.2
-Markdown==3.4.1
-MarkupSafe==2.1.2
 matplotlib==3.7.0
-matplotlib-inline==0.1.6
-mistune==2.0.5
-nbclassic==0.5.1
-nbclient==0.7.2
-nbconvert==7.2.9
-nbformat==5.7.3
-nest-asyncio==1.5.6
-notebook==6.5.2
-notebook_shim==0.2.2
 numpy==1.24.2
-packaging==23.0
 pandas==1.5.3
-pandocfilters==1.5.0
 panel==0.14.3
-param==1.12.3
-parse==1.19.0
-parso==0.8.3
-pexpect==4.8.0
-pickleshare==0.7.5
-Pillow==9.4.0
-platformdirs==3.0.0
-prometheus-client==0.16.0
-prompt-toolkit==3.0.36
-psutil==5.9.4
-ptyprocess==0.7.0
-pure-eval==0.2.2
-pycparser==2.21
-pyct==0.5.0
-pyee==8.2.2
-Pygments==2.14.0
-pyparsing==3.0.9
-pyppeteer==1.0.2
-pyquery==2.0.0
-pyrsistent==0.19.3
-python-dateutil==2.8.2
-python-dotenv==0.21.1
-python-json-logger==2.0.6
-pytz==2022.7.1
-pyviz-comms==2.2.1
-PyYAML==6.0
-pyzmq==25.0.0
-requests==2.28.2
-requests-html==0.10.0
-rfc3339-validator==0.1.4
-rfc3986-validator==0.1.1
-Send2Trash==1.8.0
-sgmllib3k==1.0.0
-six==1.16.0
-sniffio==1.3.0
-soupsieve==2.4
-stack-data==0.6.2
-terminado==0.17.1
-tinycss2==1.2.1
-tomli==2.0.1
-tornado==6.2
-tqdm==4.64.1
-traitlets==5.9.0
-typing_extensions==4.4.0
-uri-template==1.2.0
-urllib3==1.26.14
-w3lib==2.1.1
-wcwidth==0.2.6
-webcolors==1.12
-webencodings==0.5.1
-websocket-client==1.5.1
-websockets==10.4
-xyzservices==2022.9.0
 yahoo-fin==0.8.9.1
-zipp==3.13.0
