--- test_modules/tab0.py
@@ -1,39 +1,0 @@
-
-
-def get_intro():
-    text = """
-    <h1>Welcome to the Portfolio Selection Tool</h1>
-
-
-    <h4>
-    To begin, please complete the Risk Analysis Survey found in the left-hand side panel. Once you have selected your answers click on the 'submit' button.</br></br>
-
-    Once your answers are submitted a Risk Score will be calculated and you will be placed into one of four Risk Tolerance Categories. 
-    Based upon your Risk Category you will be shown details related to a portfolio of assets that best conforms to your Risk Tolerance Category.
-    </br></br>
-
-    You may naviate through the informational displays by selecting the desired tab from the menu above. The information is organized as follows:
-    </br></br>
-
-    (show a nice table listing the information available on each tab)
-    </h4>
-    <br> <br> <br> <br> <br> <br> <br> <br> <br> <br> <br> <br> <br>
-    The information provided on this website is for information and educational purposes only. 
-    <br>
-    It is not intended to be, nor should be used as, investment advice. 
-    <br>
-    Seek a duly licensed professional for investment advice.
-        """
-    
-    return text
-
-
-def get_disclaimer():
-    text = """
-    The information provided on this website is for information and educational purposes only. 
-    It is not intended to be, nor should be used as, investment advice. 
-    Seek a duly licensed professional for investment advice.
-    """
-    
-    return text
-
