--- weights1.csv
@@ -1,5 +1,0 @@
-stock,weight
-appl,0.3
-cisco,0.2
-ibm,0.4
-amzn,0.1
